+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Markdown syntax highlighting while editing notes.
+
 ## [0.2.0] - 2026-08-11
 
 ### Added